requirements.txt
python-telegram-bot[ext]
deep-translator
python-dotenv
flask